% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/back.R
\name{back}
\alias{back}
\alias{back,CppQueue-method}
\alias{back,CppVector-method}
\alias{back,CppDeque-method}
\alias{back,CppList-method}
\title{Access last element}
\usage{
back(x)
}
\arguments{
\item{x}{A CppQueue, CppVector, CppDeque, or CppList object.}
}
\value{
Returns the last element.
}
\description{
Access the last element in a container without removing it.
}
\details{
In a CppQueue, the last element is the last inserted one.
}
\examples{
q <- cpp_queue(1:4)
q
# First element: 1

back(q)
# [1] 4

l <- cpp_list(1:4)
l
# 1 2 3 4

back(l)
# [1] 4

}
\seealso{
\link{front}, \link{top}, \link{push_back}, \link{emplace_back}, \link{pop_back}.
}
