% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear.R
\name{clear}
\alias{clear}
\alias{clear,CppSet-method}
\alias{clear,CppUnorderedSet-method}
\alias{clear,CppMultiset-method}
\alias{clear,CppUnorderedMultiset-method}
\alias{clear,CppMap-method}
\alias{clear,CppUnorderedMap-method}
\alias{clear,CppMultimap-method}
\alias{clear,CppUnorderedMultimap-method}
\alias{clear,CppVector-method}
\alias{clear,CppDeque-method}
\alias{clear,CppForwardList-method}
\alias{clear,CppList-method}
\title{Clear the container}
\usage{
clear(x)
}
\arguments{
\item{x}{A CppSet, CppUnorderedSet, CppMultiset, CppUnorderedMultiset, CppMap, CppUnorderedMap, CppMultimap, CppUnorderedMultimap, CppVector, CppDeque, 
CppForwardList, or CppList object.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Remove all elements from a container by reference.
}
\examples{
l <- cpp_forward_list(4:9)
l
# 4 5 6 7 8 9

clear(l)
l
# 
empty(l)
# [1] TRUE

}
\seealso{
\link{erase}, \link{remove.}, \link{empty}.
}
