% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deque.R
\name{cpp_deque}
\alias{cpp_deque}
\title{Create deque}
\usage{
cpp_deque(x)
}
\arguments{
\item{x}{An integer, numeric, character, or logical vector.}
}
\value{
Returns a CppDeque object referencing a deque in C++.
}
\description{
Create a deque, i.e. a double-ended queue.
}
\details{
C++ deque methods implemented in this package are \link{assign}, \link{at}, \link{back}, \link{clear}, \link{emplace}, \link{emplace_back}, 
\link{emplace_front}, \link{empty}, \link{erase}, \link{front}, \link{insert}, \link{max_size}, \link{pop_back}, \link{pop_front}, \link{push_back}, 
\link{push_front}, \link{resize}, \link{shrink_to_fit}, and \link{size}. The package also adds the \link{==} and \link{[} operators and various helper 
functions (\link{print}, \link{to_r}, \link{type}).

All object-creating methods in this package begin with \code{cpp_} to avoid clashes with functions from other packages, such as \code{utils::stack} and 
\code{base::vector}.
}
\examples{
d <- cpp_deque(4:6)
d
# 4 5 6

push_back(d, 1L)
d
# 4 5 6 1

push_front(d, 2L)
d
# 2 4 5 6 1

}
\seealso{
\link{cpp_vector}, \link{cpp_forward_list}, \link{cpp_list}.
}
