% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emplace_back.R
\name{emplace_back}
\alias{emplace_back}
\alias{emplace_back,CppVector-method}
\alias{emplace_back,CppDeque-method}
\alias{emplace_back,CppList-method}
\title{Add an element to the back}
\usage{
emplace_back(x, value)
}
\arguments{
\item{x}{A CppVector, CppDeque, CppList object.}

\item{value}{A value to add to \code{x}.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Add an element to the back of a container by reference in place.
}
\details{
The emplace methods only add single elements. Use \link{insert} to add multiple elements in one call.
}
\examples{
v <- cpp_vector(4:6)
v
# 4 5 6

emplace_back(v, 12L)
v
# 4 5 6 12

}
\seealso{
\link{emplace}, \link{emplace_after}, \link{emplace_front}, \link{insert}, \link{pop_back}, \link{push_back}.
}
