% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emplace_front.R
\name{emplace_front}
\alias{emplace_front}
\alias{emplace_front,CppDeque-method}
\alias{emplace_front,CppForwardList-method}
\alias{emplace_front,CppList-method}
\title{Add an element to the front}
\usage{
emplace_front(x, value)
}
\arguments{
\item{x}{A CppDeque, CppForwardList, or CppList object.}

\item{value}{A value to add to \code{x}.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Add an element to the front of a container by reference in place.
}
\details{
The emplace methods only add single elements. Use \link{insert} to add multiple elements in one call.
}
\examples{
l <- cpp_forward_list(4:6)
l
# 4 5 6

emplace_front(l, 12L)
l
# 12 4 5 6

}
\seealso{
\link{emplace}, \link{emplace_after}, \link{emplace_back}, \link{insert}, \link{pop_front}, \link{push_front}.
}
