% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{==,CppSet,CppSet-method}
\alias{==,CppSet,CppSet-method}
\title{Check equality}
\usage{
\S4method{==}{CppSet,CppSet}(e1, e2)
}
\arguments{
\item{e1}{A \code{cppcontainers} object.}

\item{e2}{A \code{cppcontainers} object.}
}
\value{
Returns \code{TRUE}, if the containers hold the same data and \code{FALSE} otherwise.
}
\description{
Check, if two containers hold identical data.
}
\examples{
x <- set(1:10)
y <- set(1:10)
x == y
[1] TRUE

y <- set(1:11)
x == y
[1] FALSE

}
