% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erase_after.R
\name{erase_after}
\alias{erase_after}
\alias{erase_after,CppForwardList-method}
\title{Erase elements}
\usage{
erase_after(x, from, to)
}
\arguments{
\item{x}{A CppForwardList object.}

\item{from}{Index after which to delete.}

\item{to}{Index until including which to delete. Indices start at 1. The function does not perform bounds checks. Indices outside \code{x} crash the 
program.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Delete elements from a forward list by reference.
}
\examples{
l <- cpp_forward_list(4:9)
l
# 4 5 6 7 8 9

erase_after(l, 2L, 4L)
l
# 4 5 8 9

}
\seealso{
\link{clear}, \link{empty}, \link{erase}, \link{remove.}.
}
