% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/front.R
\name{front}
\alias{front}
\alias{front,CppQueue-method}
\alias{front,CppVector-method}
\alias{front,CppDeque-method}
\alias{front,CppForwardList-method}
\alias{front,CppList-method}
\title{Access first element}
\usage{
front(x)
}
\arguments{
\item{x}{A CppQueue, CppVector, CppDeque, CppForwardList, or CppList object.}
}
\value{
Returns the front element.
}
\description{
Access first element in a container without removing it.
}
\examples{
q <- cpp_queue(4:6)
q
# First element: 4

front(q)
# [1] 4

q
# First element: 4

v <- cpp_vector(c(TRUE, FALSE, FALSE))
v
# TRUE FALSE FALSE

front(v)
# [1] TRUE

}
\seealso{
\link{back}, \link{emplace_front}, \link{pop}, \link{pop_front}, \link{push_front}.
}
