% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{insert}
\alias{insert}
\title{Add elements}
\usage{
insert(x, values)
}
\arguments{
\item{x}{A \code{cppcontainers} object.}

\item{values}{Values to add to \code{x}.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Add elements to a container by reference.
}
\examples{
s <- set(1:10)
s
1 2 3 4 5 6 7 8 9 10

insert(s, 11:13)
s
1 2 3 4 5 6 7 8 9 10 11 12 13

}
