% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_after.R
\name{insert_after}
\alias{insert_after}
\alias{insert_after,CppForwardList-method}
\title{Add elements}
\usage{
insert_after(x, values, position = NULL)
}
\arguments{
\item{x}{A CppForwardList object.}

\item{values}{Values to add to \code{x}.}

\item{position}{Index behind which to insert elements.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Add elements to a forward list by reference.
}
\examples{
v <- cpp_forward_list(4:6)
v
# 4 5 6

insert_after(v, 10:11, 2)
v
# 4 5 10 11 6

}
\seealso{
\link{emplace}, \link{emplace_after}, \link{insert}, \link{insert_or_assign}.
}
