% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_factor.R
\name{load_factor}
\alias{load_factor}
\alias{load_factor,CppUnorderedSet-method}
\alias{load_factor,CppUnorderedMultiset-method}
\alias{load_factor,CppUnorderedMap-method}
\alias{load_factor,CppUnorderedMultimap-method}
\title{Get the mean number of elements per bucket}
\usage{
load_factor(x)
}
\arguments{
\item{x}{A CppUnorderedSet, CppUnorderedMultiset, CppUnorderedMap, or CppUnorderedMultimap object.}
}
\value{
Returns a numeric.
}
\description{
Get the mean number of elements per bucket.
}
\examples{
s <- cpp_unordered_set(6:9)
load_factor(s)
# [1] 0.3076923

}
\seealso{
\link{bucket_count}, \link{max_bucket_count}, \link{max_load_factor}.
}
