% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_bucket_count.R
\name{max_bucket_count}
\alias{max_bucket_count}
\alias{max_bucket_count,CppUnorderedSet-method}
\alias{max_bucket_count,CppUnorderedMultiset-method}
\alias{max_bucket_count,CppUnorderedMap-method}
\alias{max_bucket_count,CppUnorderedMultimap-method}
\title{Get the maximum number of buckets}
\usage{
max_bucket_count(x)
}
\arguments{
\item{x}{A CppUnorderedSet, CppUnorderedMultiset, CppUnorderedMap, or CppUnorderedMultimap object.}
}
\value{
Returns a numeric.
}
\description{
Obtain the maximum number of buckets the container can hold.
}
\examples{
s <- cpp_unordered_set(6:10)
max_bucket_count(s)
# [1] 1.152922e+18

}
\seealso{
\link{bucket_count}, \link{load_factor}, \link{max_load_factor}.
}
