% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_size.R
\name{max_size}
\alias{max_size}
\title{Get maximum container size}
\usage{
max_size(x)
}
\arguments{
\item{x}{A \code{cppcontainers} object.}
}
\value{
Returns a numeric.
}
\description{
Get the maximum number of elements the container can hold.
}
\examples{
s <- set(1:10)
s
1 2 3 4 5 6 7 8 9 10

max_size(s)
[1] 4.611686e+17

}
