% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge,CppSet,CppSet-method}
\alias{merge,CppSet,CppSet-method}
\title{Merge two objects}
\usage{
\S4method{merge}{CppSet,CppSet}(x, y)
}
\arguments{
\item{x}{A \code{cppcontainers} object.}

\item{y}{A \code{cppcontainers} object.}

\item{...}{Has no effect. Only included for compatibility with the generic base R function.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Merge two objects by reference.
}
\details{
\code{x} and \code{y} must be of the same type. E.g. both integer or both string. The function merges elements from \code{y} that are not in 
\code{x} into \code{x} and deletes them from \code{y}. This is done by reference.
}
\examples{
x <- set(c("Hello", "there"))
y <- set(c("Hello", "world"))

merge(x, y)
x
"Hello" "there" "world"
y
"Hello"

}
