% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_back.R
\name{pop_back}
\alias{pop_back}
\alias{pop_back,CppVector-method}
\alias{pop_back,CppDeque-method}
\alias{pop_back,CppList-method}
\title{Remove an element from the back}
\usage{
pop_back(x)
}
\arguments{
\item{x}{A CppVector, CppDeque, or CppList object.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Remove an element from the back of the container by reference.
}
\examples{
l <- cpp_list(4:6)
l
# 4 5 6

pop_back(l)
l
# 4 5

}
\seealso{
\link{back}, \link{emplace_back}, \link{pop}, \link{pop_front}, \link{push_back}.
}
