% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_front.R
\name{pop_front}
\alias{pop_front}
\alias{pop_front,CppDeque-method}
\alias{pop_front,CppForwardList-method}
\alias{pop_front,CppList-method}
\title{Remove an element from the front}
\usage{
pop_front(x)
}
\arguments{
\item{x}{A CppDeque, CppForwardList, or CppList object.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Remove an element from the front of the container by reference.
}
\examples{
d <- cpp_deque(4:6)
d
# 4 5 6

pop_front(d)
d
# 5 6

}
\seealso{
\link{emplace_front}, \link{front}, \link{pop}, \link{pop_back}, \link{push_front}.
}
