% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print,CppSet-method}
\alias{print,CppSet-method}
\title{Print container data}
\usage{
\S4method{print}{CppSet}(x, ...)
}
\arguments{
\item{x}{A \code{cppcontainers} object.}

\item{...}{An ellipsis for compatibility with the generic method. Accepts the parameters \code{n}, \code{from}, and \code{to}. See \link{to_r} for their 
effects. A difference to \link{to_r} is that their omission does not induce the function to print all elements, but to print the first 100 elements.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Print the data in a container.
}
\examples{
s <- set(10:120)

print(s)
First 100 elements: 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27 28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43 44 45 46 47 48 49 50 51 52 53 
54 55 56 57 58 59 60 61 62 63 64 65 66 67 68 69 70 71 72 73 74 75 76 77 78 79 80 81 82 83 84 85 86 87 88 89 90 91 92 93 94 95 96 97 98 99 100 101 102 
103 104 105 106 107 108 109

print(s, n = 10)
10 11 12 13 14 15 16 17 18 19

print(s, from = 42, to = 44)
42 43 44

}
