% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push_back.R
\name{push_back}
\alias{push_back}
\alias{push_back,CppVector-method}
\alias{push_back,CppDeque-method}
\alias{push_back,CppList-method}
\title{Add an element to the back}
\usage{
push_back(x, value)
}
\arguments{
\item{x}{A CppVector, CppDeque, or CppList object.}

\item{value}{A value to add to \code{x}.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Add an element to the back of a container by reference.
}
\examples{
v <- cpp_vector(4:6)
v
# 4 5 6

push_back(v, 14L)
v
# 4 5 6 14

}
\seealso{
\link{back}, \link{emplace_back}, \link{insert}, \link{pop_back}, \link{push_front}.
}
