% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push_front.R
\name{push_front}
\alias{push_front}
\alias{push_front,CppDeque-method}
\alias{push_front,CppForwardList-method}
\alias{push_front,CppList-method}
\title{Add an element to the front}
\usage{
push_front(x, value)
}
\arguments{
\item{x}{A CppDeque, CppForwardList, or CppList object.}

\item{value}{A value to add to \code{x}.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Add an element to the front of a container by reference.
}
\examples{
d <- cpp_deque(4:6)
d
# 4 5 6

push_front(d, 14L)
d
# 14 4 5 6

}
\seealso{
\link{emplace_front}, \link{front}, \link{insert}, \link{pop_front}, \link{push_back}.
}
