% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove..R
\name{remove.}
\alias{remove.}
\alias{remove.,CppForwardList-method}
\alias{remove.,CppList-method}
\title{Remove elements}
\usage{
remove.(x, value)
}
\arguments{
\item{x}{A CppForwardList or CppList object.}

\item{value}{A value to delete from \code{x}.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Remove elements from a container by reference.
}
\details{
The method ends with a dot to avoid compatibility issues with the generic \code{base::remove}.
}
\examples{
l <- cpp_forward_list(4:6)
l
# 4 5 6

remove.(l, 5L)
l
# 4 6

}
\seealso{
\link{clear}, \link{empty}, \link{erase}.
}
