% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reserve.R
\name{reserve}
\alias{reserve}
\title{Reserve bucket space and rehash}
\usage{
reserve(x, n)
}
\arguments{
\item{x}{A \code{cppcontainers} object}

\item{n}{The minimum number of elements per bucket.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Set the number of buckets to be able to hold at least \code{n} elements and rehash.
}
\examples{
s <- unordered_set(1:10)
reserve(s, 3)

}
