% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse.R
\name{reverse}
\alias{reverse}
\alias{reverse,CppForwardList-method}
\alias{reverse,CppList-method}
\title{Reverse element order}
\usage{
reverse(x)
}
\arguments{
\item{x}{A CppForwardList or CppList object.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Reverses the order of the elements in the container by reference.
}
\examples{
l <- cpp_forward_list(4:9)
l
# 4 5 6 7 8 9

reverse(l)
l
# 

}
