% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set}
\alias{set}
\title{Create set}
\usage{
set(x)
}
\arguments{
\item{x}{An integer, numeric, character, or logical vector.}
}
\value{
Returns a CppSet object referencing a set in C++.
}
\description{
Create a set. Sets are containers of unique, sorted elements.
}
\details{
Sets are associative containers. They do not provide random access through an index. I.e. \code{s[2]} does not return the second element.

C++ set methods implemented in this package are \link{empty}, \link{size}, \link{max_size}, \link{clear}, \link{insert}, \link{emplace}, \link{erase}, 
\link{merge}, \link{count}, and \link{contains}. The package also adds the \code{==} operator and various helper functions (\link{print}, \link{to_r}, 
\link{type}).
}
\examples{
s <- set(6:10)
s
6 7 8 9 10

insert(s, 4:7)
s
4 5 6 7 8 9 10

print(s, from = 6)
6 7 8 9 10

s <- set(c("world", "hello", "there"))
s
"hello" "there" "world"

erase(s, "there")
s
"hello" "world"

}
