% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort.R
\name{sort}
\alias{sort}
\alias{sort,CppForwardList-method}
\alias{sort,CppList-method}
\title{Sort elements}
\usage{
sort(x, decreasing, ...)
}
\arguments{
\item{x}{A CppForwardList or CppList object.}

\item{decreasing}{Ignored.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Sorts the elements in a container by reference.
}
\details{
\code{decreasing} and \code{...} are only included for compatibility with the generic \code{base::sort} method and have no effect.
}
\examples{
l <- cpp_forward_list(c(3, 2, 4))
l
# 3 2 4

sort(l)
l
# 2 3 4

}
\seealso{
\link{sorting}, \link{unique}.
}
