% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorting.R
\name{sorting}
\alias{sorting}
\alias{sorting,CppPriorityQueue-method}
\title{Print the sorting order}
\usage{
sorting(x)
}
\arguments{
\item{x}{An CppPriorityQueue object.}
}
\value{
Returns \code{"ascending"} or \code{"descending"}.
}
\description{
Print the sorting order of a priority queue.
}
\examples{
q <- cpp_priority_queue(4:6)
sorting(q)
# [1] "descending"

q <- cpp_priority_queue(4:6, "ascending")
sorting(q)
# [1] "ascending"

}
\seealso{
\link{cpp_priority_queue}, \link{sort}.
}
