% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_r.R
\name{to_r}
\alias{to_r}
\title{Export data to R}
\usage{
to_r(x, n = NULL, from = NULL, to = NULL)
}
\arguments{
\item{x}{A \code{cppcontainers} object.}

\item{n}{The number of elements to export. If \code{n} is positive it exports elements starting at the front of the container. If \code{n} is negative, 
it starts at the back.}

\item{from}{The value from which to start exporting. If it is not a member of \code{x}, the export starts at the subsequent value. Ignored, 
when \code{n} is specified.}

\item{to}{The value until including which to export. Ignored, when \code{n} is specified.}
}
\value{
A vector.
}
\description{
Export C++ data to an R object.
}
\examples{
s <- set(11:20)
r <- to_r(s)
r
[1] 11 12 13 14 15 16 17 18 19 20

to_r(s, n = 4)
[1] 11 12 13 14

to_r(s, n = -4)
[1] 20 19 18 17

to_r(s, from = 14)
[1] 14 15 16 17 18 19 20

to_r(s, to = 18)
[1] 11 12 13 14 15 16 17 18

to_r(s, from = 14, to = 18)
[1] 14 15 16 17 18

}
