% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top.R
\name{top}
\alias{top}
\alias{top,CppStack-method}
\alias{top,CppPriorityQueue-method}
\title{Access top element}
\usage{
top(x)
}
\arguments{
\item{x}{A CppStack or CppPriorityQueue object.}
}
\value{
Returns the top element.
}
\description{
Access the top element in a container without removing it.
}
\examples{
s <- cpp_stack(1:4)
s
# Top element: 4

top(s)
# [1] 4

s
# Top element: 4

}
\seealso{
\link{emplace}, \link{pop}, \link{push}.
}
