% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear.R
\name{clear}
\alias{clear}
\title{Clear the container}
\usage{
clear(x)
}
\arguments{
\item{x}{A \code{cppcontainers} object.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Remove all elements from the container by reference.
}
\examples{
s <- set(1:10)
s
1 2 3 4 5 6 7 8 9 10 

clear(s)
s

}
