% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contains.R
\name{contains}
\alias{contains}
\title{Check for an element}
\usage{
contains(x, values)
}
\arguments{
\item{x}{A \code{cppcontainers} object.}
}
\value{
Returns a vector of the same length as \code{values}, denoting for each value whether it is part of \code{x} (\code{TRUE}) or not 
(\code{FALSE}).
}
\description{
Check, if an element exists in the container.
}
\examples{
s <- set(1:10)
s
1 2 3 4 5 6 7 8 9 10

contains(s, 9:11)
[1]  TRUE  TRUE FALSE

}
