% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emplace.R
\name{emplace}
\alias{emplace}
\title{Add an element}
\usage{
emplace(x, value)
}
\arguments{
\item{x}{A \code{cppcontainers} object.}

\item{value}{A value to add to \code{x}.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Add an element to a container by reference.
}
\examples{
s <- set(1:10)
s
1 2 3 4 5 6 7 8 9 10

emplace(s, 11)
s
1 2 3 4 5 6 7 8 9 10 11

}
