% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty.R
\name{empty}
\alias{empty}
\title{Check emptiness}
\usage{
empty(x)
}
\arguments{
\item{x}{A \code{cppcontainers} object.}
}
\value{
Returns \code{TRUE}, if the container is empty, and \code{FALSE} otherwise.
}
\description{
Check, if a container is empty, i.e. does not contain elements.
}
\examples{
s <- set(1:10)
s
1 2 3 4 5 6 7 8 9 10

clear(s)
empty(s)
[1] TRUE

}
