% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erase.R
\name{erase}
\alias{erase}
\title{Erase elements}
\usage{
erase(x, values)
}
\arguments{
\item{x}{A \code{cppcontainers} object.}

\item{values}{A vector of values to delete from \code{x}.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Delete elements from a container by reference.
}
\examples{
s <- set(1:10)
s
1 2 3 4 5 6 7 8 9 10

erase(s, 4:6)
s
1 2 3 7 8 9 10 

}
