% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_load_factor.R
\name{max_load_factor}
\alias{max_load_factor}
\title{Get or set the maximum load factor}
\usage{
max_load_factor(x, max_load = NULL)
}
\arguments{
\item{x}{A \code{cppcontainers} object}

\item{max_load}{The containers maximum load factor, i.e. number of elements per bucket. The default \code{NULL} argument retrieves the container's 
current maximum load factor. A numeric argument sets it to the specified value.}
}
\value{
Returns the numeric current maximum load factor, if \code{max_load} is \code{NULL}. Invisibly returns \code{NULL}, if \code{max_load} is 
numeric.
}
\description{
#' Get or set the maximum load factor.
}
\examples{
s <- unordered_set(1:10)
max_load_factor(s)
[1] 1

max_load_factor(s, 3)
max_load_factor(s)
[1] 3

}
