% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_size.R
\name{max_size}
\alias{max_size}
\alias{max_size,CppSet-method}
\alias{max_size,CppUnorderedSet-method}
\alias{max_size,CppMultiset-method}
\alias{max_size,CppUnorderedMultiset-method}
\alias{max_size,CppMap-method}
\alias{max_size,CppUnorderedMap-method}
\alias{max_size,CppMultimap-method}
\alias{max_size,CppUnorderedMultimap-method}
\alias{max_size,CppVector-method}
\alias{max_size,CppDeque-method}
\alias{max_size,CppForwardList-method}
\alias{max_size,CppList-method}
\title{Get maximum container size}
\usage{
max_size(x)
}
\arguments{
\item{x}{A CppSet, CppUnorderedSet, CppMultiset, CppUnorderedMultiset, CppMap, CppUnorderedMap, CppMultimap, CppUnorderedMultimap, CppVector, CppDeque, 
CppForwardList, or CppList object.}
}
\value{
Returns a numeric.
}
\description{
Obtain the maximum number of elements the container can hold.
}
\examples{
s <- cpp_deque(4:6)
s
# 4 5 6

max_size(s)
# [1] 4.611686e+18

}
\seealso{
\link{capacity}, \link{max_bucket_count}, \link{max_load_factor}, \link{size}.
}
