% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rehash.R
\name{rehash}
\alias{rehash}
\title{Set minimum bucket count and rehash}
\usage{
rehash(x, n = 0)
}
\arguments{
\item{x}{A \code{cppcontainers} object}

\item{n}{The minimum number of buckets. A value of \code{0} forces an unconditional rehash.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Set a container's minimum bucket count and rehash.
}
\examples{
s <- unordered_set(1:10)
rehash(s)
rehash(s, 3)

}
