% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.R
\name{size}
\alias{size}
\title{Get container size}
\usage{
size(x)
}
\arguments{
\item{x}{A \code{cppcontainers} object.}
}
\value{
Returns a numeric.
}
\description{
Get the number of elements in a container.
}
\examples{
s <- set(1:10)
s
1 2 3 4 5 6 7 8 9 10

size(s)
[1] 10

}
