% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type.R
\name{type}
\alias{type}
\title{Print data type}
\usage{
type(x)
}
\arguments{
\item{x}{An object of any class defined by this package.}
}
\value{
A character value.
}
\description{
Print the data type of a container.
}
\details{
The available types are integer, double, string, and boolean.
}
